Create Multiplayer Rock Paper Scissors Game
This is a game of Rock Paper Scissors between two players, player1 and player2.
The game takes their respective inputs, deletes extra spaces and check for validity because their inputs have to be either rock, paper or scissors.
If there is a wrong input or no input at all it returns an error message, else it checks for a draw.
If there is a draw the game prints their respective choices and a tie message. If there is no tie the game then checks for the winner and prints a congratulatory message to the winner.